feat: `Repository::config_snapshot()` to access configuration values. (#331)
