+++ git-repository/tests/repository/mod.rs
@@ -1,3 +1,4 @@
+mod config;
 mod object;
 mod reference;
 mod remote;
